19.12.13

+++ app/src/main/java/com/cnm/lotto/network/LottoResponse.kt
@@ -45,7 +45,7 @@ data class LottoResponse(
             @SerializedName("winningCnt")
             val winningCnt: Int,
             @SerializedName("winningPriceByRank")
-            val winningPriceByRank: Int
+            val winningPriceByRank: Long
         )
     }
 }
